improved structure of project
every page has its own page.
