+++ src/App.css
@@ -94,9 +94,9 @@ button:hover {
 }
 
 .game-container {
-  padding: 30px 20px;
+  padding: 20px;
   max-width: 1000px;
-  margin: auto;
+  margin: 100px auto 40px; 
   text-align: center;
 }
 
@@ -177,4 +177,14 @@ button:hover {
   object-fit: contain;
   border: 2px solid #6A0DAD;
   margin-bottom: 10px;
+}
+.player-info {
+  text-align: center;
+  margin: 20px 0 10px;
+  font-size: 1.4em;
+  font-weight: bold;
+  color: #E6E6FA;
+}
+.page-content {
+  padding-top: 100px; /* espace pour que la navbar ne masque pas */
 }
